Changed the date format

+++ emergency caller/ContentView.swift
@@ -76,7 +76,7 @@ struct ContentView: View {
 
 private let itemFormatter: DateFormatter = {
     let formatter = DateFormatter()
-    formatter.dateStyle = .short
+    formatter.dateStyle = .long
     formatter.timeStyle = .long
     return formatter
 }()
